refactor: Composable, Activity를 ui 패키지로 이동
