+++ app/src/main/java/nextstep/signup/ui/MainActivity.kt
@@ -1,4 +1,4 @@
-package nextstep.signup
+package nextstep.signup.ui
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
